Phase3/llm_service/requirements.txt
kafka-python==2.0.2
openai==1.7.0
anthropic==0.8.0
python-dotenv==1.0.0
requests==2.31.0
